add missed meta and refactor saver
