+++ Assets/Plugins/Sirenix/Demos/Custom Drawers.unitypackage.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 48e08dc33330d11e9d4a1b246c52e4f6
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
