+++ Assets/Plugins/Sirenix/Demos/Editor Windows.unitypackage.meta
@@ -1,6 +1,5 @@
 fileFormatVersion: 2
-guid: 6a9057866fafbb9478628f4a65434a1e
-folderAsset: yes
+guid: ed09910c0094cb27be8f3ca264680da3
 DefaultImporter:
   externalObjects: {}
   userData:
