--- Assets/Scripts/View/UI/InternetAccess/ViewPopup.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 344414e3f70587a4d84d0ddbb5dbc522
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
